Reuse the solver to speedup symbolic execution
All constraints but the condition of a branch is the same during symbolic execution. Therefore a single solver can be created once and reused for each execution of a trace.
When solving for a branch the state of the solver is saved (push), the
branch condition is added and the solver is run. After being done with
the solver, the state of the solver is resorted by popping.

Creating and reusing the solver speeds up the process of symbolic execution.